+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - IK: warm-start parameter to ``solve()``
 - Robot model gets a ``get_link()`` function
 - Simulation gets ``set_camera_transform()`` function
+- SwingFoot type: a polynomial swing foot interpolator
 - This change log
 - ZMP support areas can now take optional contact pressure limits